refactor(review changes): update translations, move division http calls to own files, file cleaning
