+++ frontend/src/common/interfaces/base-entity.interface.ts
@@ -1,6 +1,3 @@
 export interface IBaseEntity {
   id: string;
-  createdOn: Date | string;
-  updatedOn: Date | string;
-  deletedOn: Date | string;
 }
