+++ frontend/src/common/interfaces/user.interface.ts
@@ -1,0 +1,5 @@
+import { IBaseEntity } from './base-entity.interface';
+
+export interface IUser extends IBaseEntity {
+  name: string;
+}
