--- frontend/src/common/interfaces/username.interface.ts
@@ -1,4 +1,0 @@
-export interface IUsername {
-  id: string;
-  name: string;
-}
